Created sunburstSeries file
